update standard types

+++ ProjectDevelopment/TESS/Mcal/StandardTypes.h
@@ -11,10 +11,10 @@
 typedef unsigned char uint8;
 typedef signed   char int8;
 
-typedef unsigned int  uint16;
-typedef signed   int  int16;
+typedef unsigned short int  uint16;
+typedef signed   short int  int16;
 
-typedef unsigned long uint32;
-typedef signed   long int32;
+typedef unsigned int uint32;
+typedef signed   int int32;
 
 #endif /* MCAL_STANDARDTYPES_H_ */
